worker/requirements.txt
celery
redis
requests
flask
python-dotenv
